Changed Int64 type to long long.  Added debug messages to tbb2h5.  Removed obsolete messages from dalTable.
git-svn-id: http://usg.lofar.org/svn/code/trunk/src/DAL@1419 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ implement/dalBaseTypes.h
@@ -51,7 +51,7 @@ typedef short           Int16;
 typedef unsigned int   UInt32;
 typedef int             Int32;
 typedef float         Float32;
-typedef long            Int64;
+typedef long long       Int64;
 typedef double        Float64;
 
 bool BigEndian( void );
